+++ CHANGELOG.md
@@ -11,6 +11,16 @@ This format is based on [Keep a Changelog] and this project adheres to
 ### Added
 
 - `ByteSet::maybe_contains`: a [Bloom filter] function.
+- `ByteSet::is_ascii_alphabetic`: returns `true` if [`u8::is_ascii_alphabetic`] returns `true` for all bytes.
+- `ByteSet::is_ascii_uppercase`: returns `true` if [`u8::is_ascii_uppercase`] returns `true` for all bytes.
+- `ByteSet::is_ascii_lowercase`: returns `true` if [`u8::is_ascii_lowercase`] returns `true` for all bytes.
+- `ByteSet::is_ascii_alphanumeric`: returns `true` if [`u8::is_ascii_alphanumeric`] returns `true` for all bytes.
+- `ByteSet::is_ascii_digit`: returns `true` if [`u8::is_ascii_digit`] returns `true` for all bytes.
+- `ByteSet::is_ascii_hexdigit`: returns `true` if [`u8::is_ascii_hexdigit`] returns `true` for all bytes.
+- `ByteSet::is_ascii_punctuation`: returns `true` if [`u8::is_ascii_punctuation`] returns `true` for all bytes.
+- `ByteSet::is_ascii_graphic`: returns `true` if [`u8::is_ascii_graphic`] returns `true` for all bytes.
+- `ByteSet::is_ascii_whitespace`: returns `true` if [`u8::is_ascii_whitespace`] returns `true` for all bytes.
+- `ByteSet::is_ascii_control`: returns `true` if [`u8::is_ascii_control`] returns `true` for all bytes.
 
 ## 0.1.0 - 2020-06-01
 
@@ -25,5 +35,16 @@ Initial release!
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
 [Bloom filter]: https://en.wikipedia.org/wiki/Bloom_filter
+
+[`u8::is_ascii_alphabetic`]:   https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_alphabetic
+[`u8::is_ascii_uppercase`]:    https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_uppercase
+[`u8::is_ascii_lowercase`]:    https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_lowercase
+[`u8::is_ascii_alphanumeric`]: https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_alphanumeric
+[`u8::is_ascii_digit`]:        https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_digit
+[`u8::is_ascii_hexdigit`]:     https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_hexdigit
+[`u8::is_ascii_punctuation`]:  https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_punctuation
+[`u8::is_ascii_graphic`]:      https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_graphic
+[`u8::is_ascii_whitespace`]:   https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_whitespace
+[`u8::is_ascii_control`]:      https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_control
 
 [Unreleased]: https://github.com/nvzqz/static-assertions-rs/compare/v0.1.0...HEAD